+++ Software/MATLAB/app/functions/azDLRoot.m
@@ -1,6 +1,6 @@
-function [str] = azRoot(varargin) 
-% AZROOT Helper function to locate the Azure tooling.
-% 
+function [str] = azDLRoot(varargin)
+% AZDLROOT Helper function to locate the Azure Data Lake interface
+%
 % Locate the installation of the Azure tooling to allow easier construction
 % of absolute paths to the required dependencies.